src/requirements.txt
flask==2.3.3
boto3==1.28.38
werkzeug==2.3.7
pillow
